fix: 패키지 startegy -> strategy명 오류 수정
