+++ src/main/java/gravit/code/auth/service/oauth/OAuthAndroidClientService.java
@@ -2,7 +2,7 @@ package gravit.code.auth.service.oauth;
 
 import gravit.code.auth.client.config.Auth0Props;
 import gravit.code.auth.dto.oauth.OAuthUserInfo;
-import gravit.code.auth.startegy.android.AndroidUserInfoFactory;
+import gravit.code.auth.strategy.android.AndroidUserInfoFactory;
 import gravit.code.global.exception.domain.CustomErrorCode;
 import gravit.code.global.exception.domain.RestApiException;
 import lombok.RequiredArgsConstructor;
