+++ src/main/java/gravit/code/auth/strategy/KakaoOAuthResponseStrategy.java
@@ -1,21 +1,21 @@
-package gravit.code.auth.startegy;
+package gravit.code.auth.strategy;
 
-
-import gravit.code.auth.dto.oauth.NaverUserInfo;
+import gravit.code.auth.dto.oauth.KakaoUserInfo;
 import gravit.code.auth.dto.oauth.OAuthUserInfo;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
 
 @Component
-public class NaverOAuthResponseStrategy implements OAuthResponseStrategy {
+public class KakaoOAuthResponseStrategy implements OAuthResponseStrategy {
     @Override
     public String getProviderName() {
-        return "naver";
+        return "kakao";
     }
 
     @Override
     public OAuthUserInfo createOAuthUserInfo(Map<String, Object> attributes) {
-        return new NaverUserInfo(attributes);
+        System.out.println(attributes);
+        return new KakaoUserInfo(attributes);
     }
 }
